app/requirements.txt
streamlit
python-dotenv
openai>=1.13
pillow
requests
